Added fetch all alumni data changes

+++ src/main/java/com/Alumni_LinkedIn_Profile_Searcher/repository/AlumniRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlumniRepository extends ReactiveCrudRepository<Alumni,Long> {
+public interface AlumniRepository extends ReactiveCrudRepository<Alumni,Integer> {
 
 }
